+++ core/event/service.go
@@ -18,9 +18,7 @@ type Service interface {
 type Manager interface {
 	// Emit emits events represented as a protobuf message (as described in ADR 032).
 	//
-	// Callers SHOULD assume that these events may be included in consensus. These events
-	// MUST be emitted deterministically and adding, removing or changing these events SHOULD
-	// be considered state-machine breaking.
+	// Callers SHOULD assume that these events will not be included in consensus.
 	Emit(event protoiface.MessageV1) error
 
 	// EmitKV emits an event based on an event and kv-pair attributes.
@@ -28,11 +26,4 @@ type Manager interface {
 	// These events will not be part of consensus and adding, removing or changing these events is
 	// not a state-machine breaking change.
 	EmitKV(eventType string, attrs ...Attribute) error
-
-	// EmitNonConsensus emits events represented as a protobuf message (as described in ADR 032), without
-	// including it in blockchain consensus.
-	//
-	// These events will not be part of consensus and adding, removing or changing events is
-	// not a state-machine breaking change.
-	EmitNonConsensus(event protoiface.MessageV1) error
 }