+++ types/module/module.go
@@ -312,10 +312,10 @@ func (m *Manager) RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
 }
 
 // RegisterInterfaces registers all module interface types
-func (m *Manager) RegisterInterfaces(registry registry.LegacyRegistry) {
+func (m *Manager) RegisterInterfaces(registrar registry.InterfaceRegistrar) {
 	for _, b := range m.Modules {
 		if mod, ok := b.(appmodule.HasRegisterInterfaces); ok {
-			mod.RegisterInterfaces(registry)
+			mod.RegisterInterfaces(registrar)
 		}
 	}
 }